Add AmazingFeature

+++ api/routes/books.py
@@ -1,7 +1,7 @@
 from typing import OrderedDict
 
 from fastapi import APIRouter, HTTPException, status
-from fastapi.responses import JSONResponse
+from fastapi.responses import JSONResponse, Response
 
 from api.db.schemas import Book, Genre, InMemoryDB
 
@@ -66,6 +66,11 @@ async def update_book(book_id: int, book: Book) -> Book:
 
 
 @router.delete("/{book_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_book(book_id: int) -> None:
+async def delete_book(book_id: int):
+    book = db.get_book(book_id)
+    if not book:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Book not found")
+    
     db.delete_book(book_id)
-    return JSONResponse(status_code=status.HTTP_204_NO_CONTENT, content=None)
+    print("Book deleted successfully")
+    return Response(status_code=status.HTTP_204_NO_CONTENT)
